+++ packages/gamut-labs/src/experimental/AppHeaderMobile/AppHeaderLinkMobile/index.tsx
@@ -52,7 +52,6 @@ export const AppHeaderLinkMobile: React.FC<AppHeaderLinkMobileProps> = ({
     <SeparatorOuter topSeparator={topSeparator}>
       <SeparatorInner topSeparator={topSeparator}>
         <AppHeaderLinkButtonOuter
-          data-intellimize={item.dataIntellimizeId}
           data-testid={item.dataTestId}
           href={item.href}
           onClick={(event) => action(event, item)}